feat: add WEL template

+++ index.js
@@ -49,42 +49,59 @@ const convert = file => {
     if (err) {
       throw err;
     }
+    const extname = path.extname(file);
 
-    svgo
-      .optimize(data)
-      .then(result => {
-        const svg = result.data;
+    if (extname === ".svg") {
+      svgo
+        .optimize(data)
+        .then(result => {
+          const svg = result.data;
 
-        svgtojsx(svg)
-          .then(jsx => {
-            const js = prettier.format(template(jsx));
-            const filename = file.replace(".svg", ".js");
+          svgtojsx(svg)
+            .then(jsx => {
+              const displayName = `${
+                program.prefix ? program.prefix : ""
+              }${pascalCase(path.basename(file, extname))}`;
+              const js = prettier.format(template(jsx, displayName));
+              const filename = file.replace(".svg", ".js");
 
-            fs.writeFile(filename, js, "utf-8", err => {
-              if (err) {
-                throw err;
-              }
+              fs.writeFile(filename, js, "utf-8", err => {
+                if (err) {
+                  throw err;
+                }
 
-              if (program.delete) {
-                fs.unlink(file, err => {
-                  if (err) {
-                    throw err;
-                  }
-                });
-              }
+                if (program.delete) {
+                  fs.unlink(file, err => {
+                    if (err) {
+                      throw err;
+                    }
+                  });
+                }
+              });
+            })
+            .catch(err => {
+              throw err;
             });
-          })
-          .catch(err => {
-            throw err;
-          });
-      })
-      .catch(err => {
-        console.error(err);
-      });
+        })
+        .catch(err => {
+          console.error(err);
+        });
+    }
   });
 };
 
-const template = svg => `import React from "react";export default (${svg});`;
+const pascalCase = string => {
+  return string
+    .match(/[a-z0-9]+/gi)
+    .map(word => word.charAt(0).toUpperCase() + word.substr(1).toLowerCase())
+    .join("");
+};
+
+const template = (jsx, displayName) => {
+  return `import PropTypes from "prop-types";import React from "react";import Svg from "@wel-ui/component-svg";const Icon = props => <Svg {...props}>{Icon.svg}</Svg>;
+
+  Icon.propTypes = {height: PropTypes.number,label: PropTypes.string,width: PropTypes.number};Icon.displayName = "${displayName}";Icon.svg = ${jsx};export default Icon;`;
+};
 
 program
   .arguments("<path>")
